fix frontend component usage
