--- frontend/src/components/base/Base.vue
@@ -1,7 +1,0 @@
-<script>
-
-</script>
-
-<template>
-
-</template>
